requirements.txt
# =============================
# SEAC — Sistema Experto Asistente de Compras
# Dependencias principales
# =============================

# Framework principal
fastapi==0.115.2

# Servidor ASGI (incluye uvloop, httptools, watchfiles para reload)
uvicorn[standard]==0.30.6

# Manejo de variables de entorno (.env)
python-dotenv==1.0.1

# Motor de plantillas HTML
jinja2==3.1.4

# Conexión a SQL Server (requiere ODBC Driver 17/18 en Windows)
pyodbc==5.1.0

# Validación de modelos (Pydantic v2)
pydantic==2.9.2

# Cliente HTTP (para Groq y APIs externas)
requests==2.34.0

# Servir archivos estáticos (para /static)
aiofiles==23.2.1
